src/requirements.txt
pvporcupine  == 2.2.0
sounddevice  == 0.4.6
pydub        == 0.25.1
pycaw        == 20230407
pyyaml       == 6.0
python-levenshtein == 0.20.9
vosk         == 0.3.45
torch        == 2.0.0
omegaconf    == 2.3.0
gpytranslate == 1.5.1
openai       == 0.27.4